+++ docs/demos/Split.demo.pane.tsx
@@ -56,6 +56,34 @@ export const pane: MantineDemo = {
       libraryValue: 'vertical',
     },
     {
+      prop: 'cursorVertical',
+      type: 'select',
+      data: [
+        'auto', 'default', 'none', 'context-menu', 'help', 'pointer', 'progress', 'wait',
+        'cell', 'crosshair', 'text', 'vertical-text', 'alias', 'copy', 'move', 'no-drop',
+        'not-allowed', 'grab', 'grabbing', 'all-scroll', 'col-resize', 'row-resize', 
+        'n-resize', 'e-resize', 's-resize', 'w-resize', 'ne-resize', 'nw-resize', 
+        'se-resize', 'sw-resize', 'ew-resize', 'ns-resize', 'nesw-resize', 'nwse-resize', 
+        'zoom-in', 'zoom-out'
+      ],
+      initialValue: 'col-resize',
+      libraryValue: 'col-resize',
+    },
+    {
+      prop: 'cursorHorizontal',
+      type: 'select',
+      data: [
+        'auto', 'default', 'none', 'context-menu', 'help', 'pointer', 'progress', 'wait',
+        'cell', 'crosshair', 'text', 'vertical-text', 'alias', 'copy', 'move', 'no-drop',
+        'not-allowed', 'grab', 'grabbing', 'all-scroll', 'col-resize', 'row-resize', 
+        'n-resize', 'e-resize', 's-resize', 'w-resize', 'ne-resize', 'nw-resize', 
+        'se-resize', 'sw-resize', 'ew-resize', 'ns-resize', 'nesw-resize', 'nwse-resize', 
+        'zoom-in', 'zoom-out'
+      ],
+      initialValue: 'row-resize',
+      libraryValue: 'row-resize',
+    },
+    {
       prop: 'size',
       type: 'size',
       initialValue: 'sm',